+++ Observer/Model/Observation+CsvExport.m
@@ -47,9 +47,9 @@
     NSCalendar *gregorian = [[NSCalendar alloc] initWithCalendarIdentifier:NSGregorianCalendar];
     NSInteger year = [gregorian components:NSYearCalendarUnit fromDate:self.timestamp].year;
     NSUInteger dayOfYear = [gregorian ordinalityOfUnit:NSDayCalendarUnit inUnit:NSYearCalendarUnit forDate:self.timestamp];
-    [csv appendFormat:@"%@,%@,%d,%d,%0.6f,%0.6f,%0.6f,%0.6f,WGS84",
+    [csv appendFormat:@"%@,%@,%d,%lu,%0.6f,%0.6f,%0.6f,%0.6f,WGS84",
      [AKRFormatter utcIsoStringFromDate:self.timestamp],
-     [AKRFormatter localIsoStringFromDate:self.timestamp],year,dayOfYear,
+     [AKRFormatter localIsoStringFromDate:self.timestamp],year,(unsigned long)dayOfYear,
      featureLocation.latitude, featureLocation.longitude,
      observerLocation.latitude, observerLocation.longitude];
