+++ Observer/Protocols/SProtocol.m
@@ -222,7 +222,7 @@
                             [self processProtocolJSON:json version:version];
                             break;
                         default:
-                            AKRLog(@"Unsupported version (%d) of the NPS-Protocol-Specification", version);
+                            AKRLog(@"Unsupported version (%ld) of the NPS-Protocol-Specification", (long)version);
                             break;
                     }
                 }
